+++ rules/ecmascript-6.js
@@ -61,6 +61,9 @@ module.exports = {
   // after declared
   'prefer-const': 2,
 
+  // Disallow parseInt() in favor of binary, octal, and hexadecimal literals
+  `prefer-numeric-literals`: 2,
+
   // Suggest using Reflect methods where applicable
   'prefer-reflect': 0,