+++ react/index.js
@@ -5,6 +5,7 @@ module.exports = {
 
   parserOptions: {
     ecmaFeatures: {
+      experimentalObjectRestSpread: true,
       jsx: true
     }
   },