+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `prefer-numeric-literals` rule to disallow `parseInt()` in favor of binary, octal, and hexadecimal literals
 - Add `no-restricted-properties` to not disallow certain object properties
 - Add `line-comment-position` rule to enforce position of line comments
+- Add `lines-around-directive` rule to require newlines around directives
 
 ### Changed
 - Change Gitter badge to a matching flat icon