+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `lines-around-directive` rule to require newlines around directives
 - Add `ignoreStrings`, `ignoreRegExpLiterals`, and `ignoreTemplateLiterals` options to `max-len` rule
 - Add `CallExpression` option to `indent` rule
+- Add ES7 and ES8 environments
 
 ### Changed
 - Change Gitter badge to a matching flat icon