+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `ignoreStrings`, `ignoreRegExpLiterals`, and `ignoreTemplateLiterals` options to `max-len` rule
 - Add `CallExpression` option to `indent` rule
 - Add ES7 and ES8 environments
+- Add `experimentalObjectRestSpread` parser option to React rules
 
 ### Changed
 - Change Gitter badge to a matching flat icon