+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add the `functionPrototypeMethods` option to the `wrap-iife` rule
 - Add `no-useless-return` rule to disallow redundant return statements
 - Remove deprecated `prefer-reflect` rule
+- Add `no-return-await` rule to disallow unnecessary `return await`
 
 ### Changed
 - Change Gitter badge to a matching flat icon