+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All changes to this project are documented in this file. This project adheres to
 - Change Gitter badge to a matching flat icon
 - Default to ES6 rules
 - Re-categorize `comma-dangle` rule under stylistic issues
+- Improve the description of the `comma-style` rule
 
 ## [9.1.1] - 2016-08-30
 ### Changed