+++ CHANGELOG.md
@@ -28,9 +28,10 @@ All changes to this project are documented in this file. This project adheres to
 - Default to ES6 rules
 - Re-categorize `comma-dangle` rule under stylistic issues
 - Improve the description of the `comma-style` rule
-- Update eslint dependency to ^3.10.2
+- Update `eslint` dependency to ^3.10.2
 - Enable `react/prop-types` rule with `skipUndeclared` option
 - Change to the `multiline-multiprop` option for the `react/jsx-first-prop-new-line` rule
+- Update `eslint-plugin-react` dependency to ^6.7.1
 
 ## [9.1.1] - 2016-08-30
 ### Changed