+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All changes to this project are documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Fix badge URLs in the README.md
 
 ## [10.0.0] - 2016-11-21
 ### Added